Instantiate quests when adding them to quest sets
It may be interesting for a quest to be instantiated when first added to
a quest set so it can (in the future) set itself as inactive, or change
its state earlier than just before it is started. This is motivated by
a coming commit that will add the capability for quests to save/load
configuration.

Since this code is already being used from WIP branches, the method that
adds a quest to a quest set is backwards compatible; meaning it will
verify whether the argument passed is a class or an instance, and
instantiate it in the former case.

https://phabricator.endlessm.com/T24029