Allow to trigger quest actions from GActions
This commit will organize any quest actions in a dictionary mapping a
UUID->label,callback,args . This makes it easier when we display the
quest view in a Shell banner since they will activate the same action
but with a different key argument.
The reason why a UUID is used and not simply the object ID is that
if there's an issue and the Shell popup is still displaying with a
button triggering a quest action, but the Clubhouse instance is
relaunched, then we could get a potential issue when using object IDs
(as the ID that the Shell knew could potentially match a different
action than expected).

https://phabricator.endlessm.com/T23931