tools: Add a script to conveniently build a Flatpak
If the GIT_CLONE_BRANCH variable is not specified in the envinronment,
then the script produces a manifest that replaces the "git" type source
in the manifest by a "dir" type, allowing any local changes to be built.

It also adds any extra args to be added to the flatpak-builder. e.g.
 ./build-local-flatpak.sh --instal
 will install the bundle directly in the user's Flatpak installation.

https://phabricator.endlessm.com/T23931