libquest: Instantiate the quest sets when registering them
Before, QuestSet objects were only being instantiated when the user
clicked on a QuestSetButton, but they may need to be instantiazed from
an earlier stage in order to e.g. set themselves as disabled (in the
future we want to be able to hide characters in the Clubhouse).

So this patch instantiates the QuestSet objects right when they're
registered.