Refactor how the Clubhouse quest messages are shown
This patch joins the whole UI that belongs to the quest message in one
widget (before, there were parts of the message UI that were defined in
and controlled by the main window).

This patch moves all the UI parts belonging to the quest message to its
UI file, and also simplifies how the messages are actually shown, using
a fixed position instead of using the widget's margins.