Separate characters' main images from mood ones
The characters are supposed to be shown in full-body images in the
Clubhouse's main window, and only the quest views should show the
face-only (mood) images.

This patch changes how the images for the characters are loaded and
used in the code. It expects the full-body image to be called "main"
(e.g. main.png), and loads the mood images as before, i.e. each mood
image should have the mood name (e.g. happy.png).

https://phabricator.endlessm.com/T24054