Add reactive shadows to characters
This is just a bit of eye-candy, to display a shadow under the
characters which makes them stand out more, and also changes the
shadow on mouse-over, in order to hint that the object can be
clicked.

This is likely a temporary change as it has not yet been approved
by the design team.

https://phabricator.endlessm.com/T24054