libquest: Allow QuestSets to be "notified" by their Quests
QuestSets sometimes need to be aware of changes in their Quests (e.g. in
order to set their characters as highlighted in the Clubhouse). So they
now connect to their quests' "notify" signal, meaning that when a
quest's property is changed, its QuestSet's "on_quest_property_changed"
method will be called.
Thus, subclasses of the QuestSet only have to define this method in
order to be "aware" of their quests' property changes.