Allow to show different characters in the same Quest
A quest may want to show a different character in the dialog so that it
simulates a conversation. Thus, these changes use the character_id
argument that already existed (this is actually a regression) for the
show_message and show_question methods, and includes it in the signal
emitted.
From the Clubhouse side, it will now take the mentioned character ID and
set it in the Message object. If no character is specified in the
Quest's methods mentioned above, then it will use the Quest's default
character ID.

https://phabricator.endlessm.com/T24234