sample: Don't load the sample quest unless requested
The sample quest should only be shown on the presence of a special
environment variable, but due to a recent change, it'd get visible
anyway after 10 seconds.

These changes use the mentioned variable to not even load the sample
quest at all, so there are no chances of showing it in the future.