Restructure the UI so we can introduce different pages
We need an "Inventory view" in the main window (and later also an
"Episodes view"), so this patch changes how the UI is structured, and
uses a GtkStack in the main window in order to display the needed pages.

At this moment there should still be no obvious visual differences in
the UI to the user.

https://phabricator.endlessm.com/T24077