Implement inventory page
The inventory page for now just reflects the status of the item keys
(i.e. keys with a "items.key." prefix) in the GameStateService. This
means that if an item is added (or exists already) in the GSS, then it
will be shown in the inventory view, and the opposite is true when the
item is removed.

https://phabricator.endlessm.com/T24077