libquest: Add a method to retrieve a quest object by name from the Registry
This is needed for implementing a way to run a quest directly (without
needing the user to initiate it). The name that this new method expects
has to be a class name for a quest, and the QuestSet name can also be
used as a prefix, e.g. MyQuestSetClassName.MyQuestClassName .
If the QuestSet's name is not specified, then it will search for the
first quest class (of all objects) that matches the given name.

https://phabricator.endlessm.com/T24187