Add a new action to run a quest directly
This is needed for running a quest without having the user starting it.
We need it for the "first contact" quest, where the HackUnlock app is
used and where the mentioned quest should start directly without having
the Clubhouse be shown to the user.

As a second argument to this new action, the caller can decide whether
the quest is to be run inside the Clubhouse's own quest view (False), or
in the Shell (True).

https://phabricator.endlessm.com/T24187