libquest: Allow quests to be skipped
Right now, if a quest is not available, then it will block any quests
that should come after it. However, it's good to be able to tell that
a quest can be skipped.

The first use of this feature will be for the FirstContact quest as it
belongs to Ada but should only be run directly through the Clubhouse's
action (a not available + skippable state will do this).

https://phabricator.endlessm.com/T24325