Use one single signal for showing a message or question
Instead of having separate signals for sending a message and a question
from a quest, we can simplify the logic by using just one signal and
show a message or a question depending on whether we have required
answers or not.

These changes implement the idea above but only in the bits below the
quests' API. So there's still a "show_question" method until we get rid
of it in all quests.