libquest: Ask for confirm in the show_question when no choices given
If no answer choices are passed to the show_question method, then it
will show the message with the "confirm" button.

This is very convenient as it minimizes the code needed for this common
message. In the future we will delete the choices option completely so
the show_question method will simply be "show message with the
confirmation button".

https://phabricator.endlessm.com/T24354