Don't use the Clubhouse's message dialog as a quest view
There used to be two views for running Quest: one inside the GTK window
of the Clubhouse, and one that is part of the Shell.

Now we will exclusively use the Shell quest view, so this patch stops
using the "message dialog" of the Clubhouse's window for that purpose
which means we also always remove the pop-out button from it, and keep
the close button always visible.

https://phabricator.endlessm.com/T24366