Add animation system
And connect it's "step" callback to the clubhouse page, so the
callback will be called before each frame, usually at the frame rate
of the display device.

The system works like this: Each animation has a PNG with all frames
in sequence. The sprite is loaded to a GdkPixbuf.Pixbuf, and each
frame is loaded as a subpixbuf of it, thus reusing the memory. Each
animation also has metadata with the delay of each frame. The step
callback checks when the delay is passed, and tells the animation to
advance the frame.

This commit adds the system only. At the current situation nothing
should be different, because there aren't any animations added or
assets replaced.

https://phabricator.endlessm.com/T24329