clubhouse: Style the Message
Adapt the Message layout and CSS to make it look closer to the design
specs.

The box that contains the message has a fixed height. This makes the
buttons appear in the same place regardless of the number of lines in
the message label.

The box that contains the message now fills the clubhouse width, and
has zero margin at the clubhouse top. Otherwise the drop shadows
aren't visible.

The message label now has a maximum of 5 lines, and it ellipsizes
after that, to follow the design specs.

The small x button in the message top-left corner closes the message
and is clickable even when the image is on top of it, because the
image passes the input through. The assets should be designed to have
a transparent space in that corner.

https://phabricator.endlessm.com/T24369