clubhouse: Use a "normal" type-hint normal for the window
Due to having a dock type-hint, the window was not getting focused when
opened, and this affects the visuals.
To avoid this, we use the "normal" type-hint instead, and take extra
measures to make it behave like a dock window, i.e. remove the
decoration, as well as the possibility for minimizing/maximizing/move
the window.

https://phabricator.endlessm.com/T24075