animation: Don't require a character ID for an AnimationImage
The AnimationImage doesn't have conceptually anything to do with a
character (even if it's only used for character animations for now).
The character ID in the AnimationImage was also only used for correctly
replacing the animation in the AnimationSystem, and since this should
be unique to each AnimationImage, we can use the object's ID instead.

Thus, this patch drops the requirement for an ID at all.