animation: Add Animator class
The Animator class is similar to the AnimatedImage class, i.e. allows to
load several animations, but instead of being a GtkImage-based widget,
it is given the target image in which the animations should be painted.

Thus, this patch also replaces the AnimatedImage's core by the Animator,
but the real reason for this patch is that this way we can easily
animate image objects set up in the UI builder.

https://phabricator.endlessm.com/T24460