quests: Bring the sample quest back
The sample quest has been broken ever since the animation support was
added.
These changes bring it back (since it's very convenient to be able to
test all functionalities in one place), but since it required adding
a couple of animations, this patch also changes the sample quest's
character to be one of the Hack's daemons.

https://phabricator.endlessm.com/T24597