libquest: Add a "highlighted" property to the QuestSet
The QuestSet class has used the "nudge" signal to set its character as
highlighted, but even though this is conceptually correct, it has the
problem that we may need to set the QuestSet as highlighted right from
its constructor (when nothing is connecting yet to its signals).

So this patch adds the "highlighted" property to be able to check at
any time if the QuestSet intended to have its character highlighted.
In a later commit we will remove the "nudge" signal.

https://phabricator.endlessm.com/T24507