clubhouse: Set SuggestingOpen on start up
After the quest sets are loaded, the Clubhouse will check if there are
any that have the "highlighted" property set on them, and set the
SuggestingOpen property accordingly, which will effectively set the
Clubhouse button to the highlighted state.

This is needed when the Shell checks the Clubhouse's SuggestingOpen
property without having the Clubhouse fully shown, so it may highlight
the button when the desktop session starts.

https://phabricator.endlessm.com/T24507