libquest: Run the quest loop at a higher frequency
Sometimes, when clicking a button in the Shell quest view, the feedback
is not immediate. This is mainly because quests process their steps
every one second only, but this can be agravated if the steps take some
more time on their own.

Ideally we should call the next loop step when receiving the response
from the Shell, but for now let's just decrease the time slept between
each loop iteration.

https://phabricator.endlessm.com/T24592