clubhouse: Close the message if the window is closed by a quest set
When a quest is running, clicking on the same character that offered the
quest will just hide the Clubhouse window. However, if before we click
that character we had clicked in another (meaning a message is being
shown in the Clubhouse), then upon reopening the window, that message
would show as empty.

This patch hides the message when the window is closed in the mentioned
use-case.

https://phabricator.endlessm.com/T24696