clubhouse: Check if a quest is current before closing its notification
When a quest is finished, we send a request for the Shell to dismiss the
quest banner, however, at this point a new quest may already be running,
meaning we could end up dismissing the new quest's banner.

This patch avoids that by checking if there's no new quest before
dimissing the banner.

https://phabricator.endlessm.com/T24698