clubhouse: Focus the window when opened from a show-page action
Like when it's normally opened, we should also make sure the window is
focused when opening it from the "show-page" action. Thus, e.g. when the
user clicks in the "Show me" button of a quest item notification, the
window will be focused.

For a thorough explanation of the changes, see the previous commmit.

https://phabricator.endlessm.com/T24717