libquest, clubhouse: Highlight character when a quest ends
Add a new "dismissed" signal to the quest. It will be emitted when the
quest is dismissed, either completed or cancelled.

Connect the QuestSet to the dismissed signal of its quests. This is to
check if there is a next quest available. And If there is, nudge the
QuestSet to attract user attention.

https://phabricator.endlessm.com/T24863