corruption: Don't abort until app has quit for >1 s
This is a shot in the dark at fixing the corruption quest bug. I believe
it happened once that the app_is_running check occurred at the exact
moment that the hackdex app was restarting, and therefore went to the
abort step. This adds a "check abort" step that checks again in the next
iteration whether the app is still not running. If the app is running
again, it goes back to the "check goal" step. If the app is still not
running after a few checks, it moves on to the "abort" step.

https://phabricator.endlessm.com/T24870