clubhouse: Don't pulse shell button in certain conditions
The pulsing icon is intended to attract the user to open the clubhouse
when a QuestSet gets highlighted.

When a QuestSet gets highlighted and the clubhouse is open, closing
the clubhouse should not show the button pulsing.

When the highlight of a QuestSet is removed, the icon should not stop
pulsating as there could be other QuestSets highlighted.

https://phabricator.endlessm.com/T24863