quest: Gracefully handle exceptions in quest steps
If the same exception is raised 10 times in a row from the same quest
step, then we stop the quest. Otherwise we keep going. This is to allow
a chance to recover from transient failures, such as the game state
service being momentarily unavailable. Currently such failures would
knock the quest out of its main loop immediately.

https://phabricator.endlessm.com/T24893