libquest: Remove property from Quest and update QuestSet highlight logic
The Quest highlighted property that was confusing, because the
QuestSet has a property with same name. Also, it was added in 68cddc08
for a special case only. And is not needed anymore, because now we
highlight the QuestSet any time there is a quest available.

This allows us to simplify the logic to update the highlighted state
of QuestSets, and reuse the callback for quest dismissed.

- Update highlighted at startup, if there is a quest available.

- Do the above also after a quest is dismissed (completed or aborted).

- When a quest of the set becomes avaiable, AND is the next available
  quest, highlight.

This commit also rearranges logs to make them more useful:

- Move info log from QuestSet get_next_quest() to
  _update_highlighted(), because get_next_quest() is called multiple
  times for checks and becomes less relevant.

- Log "turning QuestSet visible" only when it actually happens.

https://phabricator.endlessm.com/T24863