libquest, utils: Show message with more info from spreadsheets
A message can now be displayed with a different character, mood, or
sound coming directly from the spreadsheet. This allow script
writters to define all in one place.

The only mandatory parameter for `Quest.show_message()` is now
`info_id`, an identifier for the message information in the
spreadsheet. This way, scripts can stop using the QS helper.

https://phabricator.endlessm.com/T24916