animation, clubhouse: Add alternative path for loading character animations
Animator and AnimationImage now receive a subpath that will be
searched inside the app dir and potentially overriden by an asset with
the same name in an alternative dir.

Allows testing and adjusting animations with a default installation,
without needing to rebuild and reinstall the app.

This also removes the need to load the paths for moods in the
Character construction.

https://phabricator.endlessm.com/T24968