clubhouse: Show quest popup when clicking the quest's character
Instead of just hiding the Clubhouse window when a character with a
running quest is clicked, show the last message sent from the quest too.
This new behavior comes as a complement to the new Shell's quest banner
approach when its close button is clicked: instead of stopping the
quest, it simply informs the Clubhouse that the banner has been
dismissed, and the quest stays running in the background. Thus, when
clicking the quest's character, it will show the previously shown popup
again.

https://phabricator.endlessm.com/T24641