libquest: Add set_to_background and set_to_foreground Quest methods
These methods aren't doing anything yet, but they will be called when a
quest view is dismissed, or the quest is brought back to the quest view.
They can be overridden by Quest subclasses, in order to e.g. stop quests
right away, or emit sounds, etc.

This comes after the changes to the Shell's quest view banner, which no
longer stops the quest when it's closed, but just informs the Clubhouse
about that.

https://phabricator.endlessm.com/T24977