libquest: Use a stop timeout when a quest is set to background
These changes modify the default set_to_background/foreground methods,
to set up a timeout when the quest is set to background, and to reset it
when it's brought back to the foreground.

The Quest's new property "stop_timeout" can also be overridden in order
to set up a different timeout for a quest (or to ignore it completely,
if set to -1).

https://phabricator.endlessm.com/T24977