clubhouse: Don't use quest answers system in the message popup
The in-Clubhouse message popup was using the quest's answer system from
the times when it was used to display the actual quests' messages.
This patch removes that use, because it may interfere with the ongoing
quests, now that we keep quests running in the background.

We need this change particularly, because we'll use a message popup to
get the confirmation from the user whether to foreground a task, and if
we keep using the quest's answer system for this, we'll lose the ongoing
task's answers.