libquest: Add support for a Quest's continue message
Now that quest can keep running in the background, we want the related
quest's character to ask the user for confirmation whether to continue
or to stop that quest.

For that, these changes add a continue_message property to the Quest
class (that can easily be changed to modify just the message), and a
get_continue_info method too (that returns the message and
continue/stop labels).

https://phabricator.endlessm.com/T24638