clubhouse: Initialize a variable where it's needed in the ClubhousePage
The ClubhousePage was getting the next quest for a quest_set when its
charcter is clicked. However, if the character already has a quest
running in the background, then it won't need to know the next quest.
Thus, this patch moves the call to "get_next_quest" to a later point
where it's needed, hopefully saving some processing.