Allow to override the quests' question accept/reject labels
When a user clicks a character that has a quest to be offered, a message
is shown in the message popup with the accept/reject options. Just like
the message, those options' labels should also be given by the quests
when desired.

This patch allows that by using the new accept/reject_label properties
of the Quest class, and those can also be set up directly from the
quest's string file by using the QUESTNAME_QUEST_ACCEPT/REJECT IDs.

https://phabricator.endlessm.com/T24424