libquest: Allow to skip the Quest's prefix when retrieving strings
The strings in the quests' catalog always have a prefix that matches
their quest class name (in capitals). Thus, we are able to avoid having
to always include the prefix for the string ID (when showing a message
from a quest) by just using the class' name as the prefix automatically.

This patch allows that by default (since we will change the quests code
soon), but falls back to checking the given string ID without including
the prefix when the former is not found in the catalog.

https://phabricator.endlessm.com/T24424