libquest: Set the quests' initial message automatically if needed
The initial question was consistently given by just using the class'
name+"_QUESTION" as the string ID from the quests' string catalog,
so it's possible to automate this and simplify the quests' code.

This patch does that by allowing the initial_msg argument of the
Quest class constructor to be None, in which case it will try to
retrieve the string using the name convention mentioned above.

https://phabricator.endlessm.com/T24424