system: Don't log when checking if an app is running
The Desktop.app_is_running() function is used very often by the quests
and was logging every time a quest was not running. Since this method
is just a simple check, and not an attempt to use the app, we should not
log such a message as it clutters the output quite a lot.