clubhouse: Keep the app as an instance var in the ClubhousePage
Instead of always getting the Gtk Application instance by using the
Window where associated with the ClubhousePage, these changes keep the
Application as an instance variable which makes its use simpler in
terms of the code and also safer if the window gets disassociated with
the application. This will soon be the case when we start removing the
window from the application when the former is hidden, in order to more
directly control the application's inactivity (i.e. to let it quit by
itself if it's inactive).

https://phabricator.endlessm.com/T24714