libquest: Fix getting hints for messages
The function that gets the hints automatically from the message ID was
broken after the quests started using the ID without the Quest's name as
a prefix. This patch fixes this issue by accounting for this new form,
adding the prefix automatically, and falling back to the given ID if
needed.

If no hints are found at all, then a warning is logged and no hint
button is shown.

https://phabricator.endlessm.com/T24424