Move details about how a quest is run into a Quest class
The ClubhousePage class was the one managing how a quest was run
(using a thread). However, we're changing this approach to one where
no threading is needed; thus, in order to have both approaches at
the same time until we port all quests to the new way, this patch
moves the actual implementation of how a quest is run into the Quest
class itself.

https://phabricator.endlessm.com/T24699