clubhouse: Change InventoryItem text when clicked
Display a help text in the label for 5 seconds. This is to tell the
user what to do with the key. The text is different for used and
unused keys.

Enlarge the item height to have enough space for 3 lines in the
label. This prevent the inventory rows to resize when an item is
clicked.

Set properties to the label to expand multiple lines while keeping its
width limited.

Style the active state, and add a dim effect to the icon when its item
is active.

https://phabricator.endlessm.com/T25188