libquest: Add support for running non-polling quests
Running a Quest has been so far implement by continuously calling
the same quest method every 100ms. While this approach allows quests
to implement any behavior with basic functions, the code can also
become harder to maintain, besides the implications of continuously
polling (possible CPU impact, not responsing to events until the
next loop cycle, etc.), or running in a thread (possible concurrency
issues).

These changes make use of the asyncio Python library, particularly
its GLib implementation, to provide a way for quests to be written
without any polling. Quests can now call for a pause, or wait for
user confirmation without its method being repeatedly called just
for iterating a time counter.
In a later commit, functions for waiting for an app launch or an
app property change will be added.

This patch still keeps the mentioned polling method available,
distinguishing between which one to be used by the existence of
either the method "step_begin" (in which case it will use the new
way), falling back to the old way.
When all quests are ported to the new way, we will delete the
polling implementation.

https://phabricator.endlessm.com/T24699