tools: Add new setup-lint script
This script sets up Flake8 in a Python virtual environment and using
the Platform runtime. This way it's not only easy to set up for all
developers, but it also ensures that the versions can be easily kept
the same and up to date, and should also work when we set up lint
checks from the CI system.

https://phabricator.endlessm.com/T25217