system, libquest: Add API to handle when the app window is focused
Quests can already detect if the app is running. But its window could
be under another app window. With this change, quests can detect if
the app is not in the foreground, and ask the user to raise it.

The nomenclature is hard because the system Desktop already has a
method `focus_app()` which focuses the icon inside the icon grid. The
names come from the Shell. The Shell attribute `FocusedApp` used in
this commit through D-Bus is for the app window, not the app icon.

https://phabricator.endlessm.com/T24620