libquest: Remove delay argument from the set_next_step function
The set_next_step function was taking the delay argument in order to call
the next step function after waiting for a while. However, the same is
possible by using the "pause" method in the step function returning the
next step, and it's possibly more clear to use it than returning the
next step with a delay, so this patch removes this possibility.