libquest: Add wait_for_one Quest method
This method is passed a list of AsyncActions and will block until one
of them is done. This is useful for quests that need to wait for more
than one event, e.g. waiting for the user to confirm a message, while
responding to changes in an app.

The method is implemented in the _RunContext helper class, and its
wait_for_action method's implementation is replaced by a wait_for_one
call.

https://phabricator.endlessm.com/T25214