libquest: Add Quest's connect_app_js_prop_changes method
This method, unlike the existing wait_for_app_js_prop_changed will
not block until there's a property change, but instead returns an
AsyncAction object that can be used for waiting for the change.

This is useful if a quest needs to react to apps' property changes
but still not block.

https://phabricator.endlessm.com/T25214