quests: Make FizzicsIntro react to level changes without beating them
The FizzicsIntro was only reacting to beating levels, i.e. when the
levelSuccess property changed. However, if the level was changed by
a different means, e.g. by just advancing to the next level in the UI,
then the quest wouldn't detect that.
To fix that, this patch makes the quest listen to level changes, in
addition to beating the level.

https://phabricator.endlessm.com/T25257