libquest: Fix Quest.with_app_launched decorator's failure callback call
The failure (otherwise) callback was being called with all the arguments
that were passed to the wrapped function, and this can be problematic
since the failure function may not expect those arguments.

To fix this, this patch ensures that only the first argument is passed,
since we assume that it needs to be an instance method (not a static or
class one).

https://phabricator.endlessm.com/T25282