libquest: Monitor app's life in the "with_app_launched" decorator
The mentioned decorator only checks if the application is running
before running the wrapped function, however, that function can block
for a while (waiting on user inpout of a change in the system).
Thus, these changes make the decorator call the passed "failure
function" even if the app stopped running while the wrapped function
is executing. Such behavior is what's expected from the steps before
the port to the non-polling approach.

https://phabricator.endlessm.com/T25282