Fix use of get_object_property and get_js_property
The commit cf7b1afb was a mistake that sneaked in through the review
and stops the App.get_object_property from throwing exceptions. This
change was however aimed for App.get_js_property (which uses
get_object_property within). As a result, it may return unexpected
values in quests that are using it.

This patch reverts the changes to the get_object_property, and makes
the get_js_property not raise exceptions instead. It also adapts the
code using both functions, so both the newly ported quests and legacy
ones keep working as expected.

https://phabricator.endlessm.com/T25214