system: Change how the App notifies of its properties changes
Due to a bug in Clippy, it may continuously notify of property changes
even if the value didn't actually change. It was not noticeable for
Clubhouse users due to how the quests were implemented, but with the new
(non-polling) approach, the issue is exposed, and even though it is
fixed in Clippy already, some users may still experiment this problem
until they upgrade.

Thus, this patch fixes that issue in the App class by checking if its
properties' values have actually changed before calling the connected
callback.

https://phabricator.endlessm.com/T25276