libquest: Add Quest method to get the current "confirm action"
Even thought we can already verify whether a user has confirmed a message,
this has always been tied to showing a message as well, and this is undesired
when waiting for the confirmation together with other changes (e.g. settings'
changes), as we don't want to show a new message every time the wait is
restarted.

Thus, with this new method, one can show a confirmation message just once,
and then connect to the user confirmation action and any other changes
repeatedly if needed.
An example illustrates this better:

  show_confirm_message(...)
  while not settings_get_value(...) == 'expected':
    wait_for_one(get_confirm_action, other_async_action, other_async_action)

If we had to use the return value of show_confirm_message() above, then it'd
be cancelled if another async_action was resolved, and we'd have to show the
same message again just to get a new object.

https://phabricator.endlessm.com/T25214