libquest: Don't return broken AsyncActions when cancelled
If the quest has been cancelled, then the _RunContext was returning any
new AsyncAction objects without setting their run_context or future.
The original idea was that if the quest has been cancelled, then these
members shouldn't be used anyway, but the fact is that if we have
consecutive calls for methods that block and return AsyncAction, they
will still be called even if the quest has been cancelled, i.e. if a
quest calls e.g.:
    self.show_confirm_message(STR_ID).wait()
    self.show_confirm_message(STR_ID).wait()

and the quest is cancelled while the first wait_confirm call is pending,
then the second wait_confirm call would use an AsyncAction object that
has no run_context nor future inside of it, meaning that if the method
calls "wait" on it, the assert exception when checking for the
RunContext would be reached.

https://phabricator.endlessm.com/T25276