clubhouse: Play sound when item accepted
This changes the item-accept-answer action to take one boolean parameter
which tells whether to show the item in the inventory screen, or not.
In either case a sound is played. If the parameter is True, then
additionally the inventory is opened as if the show-page('inventory')
action had been called.

https://phabricator.endlessm.com/T25324