utils: Strip the text coming from QuestStringCatalog
The Shell is presenting empty dialogs if if there are many newlines at
the end of a Quest message's text. While this may be a problem in how
the Shell's Clubhouse component splits the text, there's also no
point in sending text with leading/ending white spaces.

Thus, in order to prevent this issue, this patch strips the heading
and trailing spaces off of the text coming from the quests' strings.