episodes: Make sure the poster window cannot be dragged
Even though the dialog of the poster window was not draggable, the
maximized window (transparent background) was still able to be
dragged by using Super+mouse move.
This patch ensures that cannot be done by setting the GdkWindow's
functions to only allow "close" (it also refactors a bit the code
for doing the same to the dialog).

https://phabricator.endlessm.com/T25123