quests: episode1: osintro: Remove Saniel intro sound
The sound is in the spreadsheet. Because this sound is tagged as
"bg" in the hack-sound-server, this sound should be moved to the
BG sound column. Otherwise if you step to the next dialog while
the saniel-intro sound is playing this will paused and weirdly
resumed back when the last bg sound finishes.

https://phabricator.endlessm.com/T25111

https://phabricator.endlessm.com/T25111