Add new apps module with LightSpeed app
This new module is for classes that represent specific apps used by
the quests which should abstract some common operations on them, e.g.
set the level, get the score, and others whose code is currently
repeated across quests.

For now it has only the LightSpeed app, but it should have more soon.