quests: Port PowerUps1 quest to the non-polling approach
This quest didn't check any properties in the app (it's simply a
skeleton waiting for the user debug action to occur), so I have made
this new version wait for a dummy property on the app. This way we
have the same effect as before but it's easier to spot which methods
need to be changed (as the errors are shown on the log).

https://phabricator.endlessm.com/T25404