libquest: Fix multiple use of Quest.with_app_launched
If the Quest.with_app_launched decorated is used, and its app is closed
then the second time the same decorated function is called (i.e. if the
quest is restarted), the decorator will always act as if the app had been
called. This happened due to a scope issue related to how the decorator's
local variables were being used; particularly the local variable that was
setting whether the app had been quit was declared in the decorator's
scope, not in the wrapper's scope, and thus was never reset the second
time that the decorator was used.

This patch moves the apps (and a helper function) declaration into the
wrapper, which fixes this issue.

https://phabricator.endlessm.com/T25214