libquest: Use a timeout only when the quest is in the background
Instead of using a timeout per step, the timeout for cancelling a
running quest should be set up only when it enters the background,
and reset if the quest gets back into the foreground.

This used to be the behavior with the legacy API but was not honored
when the new API was developed.

https://phabricator.endlessm.com/T25497